refactor(ui): finish impeccable audit residuals for color-safe status + mono headers
Push the audit score from Good toward Excellent by closing the last
three residuals:

- Extract a shared ConnectionStatusDot widget that conveys state by
  shape (filled dot = connected, ring = connecting/disconnected) plus a
  tooltip label, never color alone. Wire it into the Hosts list and the
  terminal tab strip, replacing the color-only 8x8 dot there.
- Bump three low-alpha onSurface text spots (key type, snippet command,
  folder label) toward AA legibility.
- Convert the three upgrade_screen section headers to lowercase mono to
  match the documented casing register.
